Refactored the ParticleSystem renderer to a system
