+++ src/GameModel.cpp
@@ -109,14 +109,13 @@ void GameModel::initialize()
         m_level->getBackgroundSize(),
         math::Point2f(0.0f, 0.0f));
 
-    m_rendererParticleSystem = std::make_unique<renderers::ParticleSystem>();
-
     m_rendererHUD = std::make_unique<renderers::HUD>();
     m_rendererStatus = std::make_unique<renderers::GameStatus>();
 
     m_sysRendererSprite = std::make_unique<systems::RendererSprite>();
     m_sysRendererAnimatedSprite = std::make_unique<systems::RendererAnimatedSprite>();
     m_sysRendererSarsCov2 = std::make_unique<systems::RendererVirus>();
+    m_sysRendererParticleSystem = std::make_unique<systems::RendererParticleSystem>();
 
     m_bullets.clear();
     m_powerups.clear();
@@ -258,14 +257,13 @@ void GameModel::render(sf::RenderTarget& renderTarget, const std::chrono::micros
 {
     renderTarget.clear(sf::Color::Black);
     m_rendererBackground->render(renderTarget);
-
-    m_rendererParticleSystem->render(m_sysParticle, renderTarget);
     m_rendererHUD->render(m_remainingNanoBots + 1, m_timePlayed, m_virusesKilled, renderTarget);
     m_rendererStatus->render(renderTarget);
 
     m_sysRendererAnimatedSprite->update(elapsedTime, renderTarget);
     m_sysRendererSarsCov2->update(elapsedTime, renderTarget);
     m_sysRendererSprite->update(elapsedTime, renderTarget);
+    m_sysRendererParticleSystem->update(m_sysParticle, renderTarget);
 }
 
 // --------------------------------------------------------------
@@ -531,6 +529,7 @@ void GameModel::addEntity(std::shared_ptr<entities::Entity> entity)
     m_sysRendererSprite->addEntity(entity);
     m_sysRendererAnimatedSprite->addEntity(entity);
     m_sysRendererSarsCov2->addEntity(entity);
+    // NOTE: Particle system renderer does not have entities added to it, it is it's own separate thing
 }
 
 // --------------------------------------------------------------
@@ -553,6 +552,7 @@ void GameModel::removeEntity(entities::Entity::IdType entityId)
     m_sysRendererSprite->removeEntity(entityId);
     m_sysRendererAnimatedSprite->removeEntity(entityId);
     m_sysRendererSarsCov2->removeEntity(entityId);
+    // NOTE: Particle system renderer does not have entities added to it, it is it's own separate thing
 }
 
 // --------------------------------------------------------------
