+++ src/systems/RendererParticleSystem.cpp
@@ -20,26 +20,27 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
 
-#pragma once
+#include "RendererParticleSystem.hpp"
 
-#include "systems/ParticleSystem.hpp"
-
-#include <SFML/Graphics.hpp>
-#include <memory>
-
-namespace renderers
+namespace systems
 {
     // --------------------------------------------------------------
     //
-    // This knows how to render the particles in a particle system.
-    // Turns out to be prety simple.
+    // This class is a friend to the renderers::ParticleSystem so that
+    // it has access to private data, without having to expose it
+    // to everyone else.
     //
     // --------------------------------------------------------------
-    class ParticleSystem
+    void RendererParticleSystem::update(systems::ParticleSystem& ps, sf::RenderTarget& renderTarget)
     {
-      public:
-        void render(systems::ParticleSystem& ps, sf::RenderTarget& renderTarget);
+        for (decltype(ps.m_particleCount) p = 0; p < ps.m_particleCount; p++)
+        {
+            auto& particle = ps.m_inUse[p];
+            particle->sprite.setPosition(particle->center);
+            particle->sprite.setRotation(particle->rotation);
 
-      private:
-    };
-} // namespace renderers
+            renderTarget.draw(particle->sprite);
+        }
+    }
+
+} // namespace systems
